refac: seperate test file
