+++ pyproject.toml
@@ -39,11 +39,12 @@ pylint = "^2.17.4"
 jinja2 = "^3.1.2"
 websockets = "^11.0.3"
 pandas = "^2.0.1"
+pytest-it = "^0.1.4"
 
 [tool.taskipy.tasks]
 local = "API_ENV=local uvicorn claon_admin.main:app --host 0.0.0.0 --port 8000 --reload"
 prod = "API_ENV=prod uvicorn claon_admin.main:app --host 0.0.0.0 --port 8000 --reload"
-test = "API_ENV=test python3 -m pytest tests -vv"
+test = "API_ENV=test python3 -m pytest tests --it"
 lint = "pylint --rcfile=.pylintrc --disable=R claon_admin"
 testCoverage = "API_ENV=test python3 -m pytest --cov-config=.coveragerc --cov=claon_admin/ --cov-report=xml"
